Move treeVisualization to its own file
